Added support for pushing packages to Azure nuget feed
The solution uses Sleet which is a static package feed utility
written by one of the NuGet team member at https://github.com/emgarten/Sleet.

It allows initializing the static JSON files for a NuGet v3
feed and also updating the relevant indexes and files when pushing
new packages.

The proper environment variables (or MSBuild command-line properties)
need to be set in order for the push to succeed.