Improve the cross-targeting checks and defaults
Leverage NuGet's build-in support for targets that are only imported
on cross-targeting builds (`buildCrossTargeting` folder) and default
values there.

Full support for 15.3+ Active Debug Framework is now included, which
allows to switch the target VS to build for easily and from within the
IDE.

The new '*' value allows to skip importing the ActiveDebugFramework
value from the cross-targeting targets, which may be useful in some
scenarios.

Dev always trumps ActiveDebugFramework, though.