Allow creating point-in-time releases without branching
We want to be able to trivially push a new version to nuget.org
(eventually) whenever a tag is created. However, we don't want
to consider tags for versioning when there are extra commits on
top. In those cases, we want to fallback to whatever the GitInfo
file contains.

This allows releasing quickly from master/dev without having to
branch, and also allows master to continue with a different
numbering (i.e. vNext) without having to change the GitInfo file
right after tagging.

So this is accomplished in two ways:
* First, we clear the GitBaseVersion if it was sourced from a tag
  but we found commits on top of it (GitCommits != 0)
* Second, we leave PackageVersion untouched both if there are no
  pre-release labels (i.e. versioned "full" release branch) as well
  as if there is a tag (whatever labels it may have).