Added ValidateVsixReferencedAssemblies task/target
Which verifies if the vsix contains assembly references that
might not be supported by the target VS version. e.g. An extension
targeting dev14 and referencing VS nuget packages with version 15.0

The target is scheduled via ValidateVsixPartsDependsOn (built-in
in the MS.VSSDK) which it's only executed if the VSIX container is
being created ('$(CreateVsixContainer)' == 'true')

By default "Microsoft;VisualStudio" string contains based references are evaluated.
But they can be customized via the ReferencedAssembliesToValidate or
ExcludeValidateVsixReferencedAssembly items.