Make versioning more useful for local builds and CI
When building locally, always append 1 to the determined GitCommits.
This means that while you're working on a new fix, your local package
version matches what it will be when built on CI (typically).

We drop the -build{number} suffix which makes every package a prerelease
one, and instead adopt a simpler numbering that can be pushed to nuget.org
as-is via a RD as needed when we're ready to do so.