Rename/split BindingRedirectAllVersions into BindingDefaultFrom and BindingDefaultTo
This makes it more explicit and intuitive on what those values mean. The previous
property affected only the upper (to) limit of the redirected 'oldVersion', so the
new properties reflect that more effectively.

Also allowing to change the default lower bound is simpler for cases where you want
to default to something other than 0.0.0.0.