Rename BindingRedirected -> BindingRedirect, fixup wrong condition
We were conditioning the generation of the pkgdef to a property that doesn't
exist anymore (the fusion expression). Also, if the `@(BindingRedirect)` is
populated more dynamically, statically calculating the property to detemrine
if generation should happen won't work.

So instead, just condition the main target `BindingRedirects` to the existence
of `@(BindingRedirect)` items at the time the target is about to be run.