Update build script, re-map VS versions
Extensions should never reference a newer target framework than net46
in VS2017, as documented in https://github.com/github/VisualStudio/issues/1849#issuecomment-411570902.

Therefore, we redesigned the mapping so it's now:

| TargetFramework | Visual Studio |
|-------|------|
| net45 | 2013 |
| net452| 2015 |
| net46 | 2017 |
| net47 | 2019 |

Switch to using Microsoft.Build.Locator instead of hacking MSBuild to
force it to resolve properly. We now run the tests against the same
MSBuild that built the tests.